feat: change precision data type to uint32

+++ contracts/libraries/SSVStorageEB.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity 0.8.24;
 
-uint64 constant VUNITS_PRECISION = 10_000;
+uint32 constant VUNITS_PRECISION = 10_000;
 uint256 constant MAX_EB_PER_VALIDATOR = 2048 ether;
 uint256 constant DEFAULT_EB_PER_VALIDATOR = 32 ether;
